chore: uint base types and skeleton for ssz (#34)
* chore: uint base types and skeleton for ssz

* some fix - wip

* fix tests

* fix silent truncation of floats

* add max method

* fix constructor

* Update src/lean_spec/types/uint.py


* add pydantic checks

* fix some operations and tests

* fix max method

* fix max

* add test for max method

---------

+++ src/lean_spec/subspecs/chain/config.py
@@ -12,16 +12,16 @@ from lean_spec.types import BasisPoint, Uint64
 
 # --- Time Parameters ---
 
-SLOT_DURATION_MS: Final = 4000
+SLOT_DURATION_MS: Final = Uint64(4000)
 """The fixed duration of a single slot in milliseconds."""
 
-SECONDS_PER_SLOT: Final = SLOT_DURATION_MS // 1000
+SECONDS_PER_SLOT: Final = SLOT_DURATION_MS // Uint64(1000)
 """The fixed duration of a single slot in seconds."""
 
-JUSTIFICATION_LOOKBACK_SLOTS: Final = 3
+JUSTIFICATION_LOOKBACK_SLOTS: Final = Uint64(3)
 """The number of slots to lookback for justification."""
 
-PROPOSER_REORG_CUTOFF_BPS: Final = 2500
+PROPOSER_REORG_CUTOFF_BPS: Final = Uint64(2500)
 """
 The deadline within a slot (in basis points) for a proposer to publish a
 block.
@@ -31,7 +31,7 @@ Honest validators may re-org blocks published after this cutoff.
 (2500 bps = 25% of slot duration).
 """
 
-VOTE_DUE_BPS: Final = 5000
+VOTE_DUE_BPS: Final = Uint64(5000)
 """
 The deadline within a slot (in basis points) by which validators must
 submit their votes.
@@ -39,7 +39,7 @@ submit their votes.
 (5000 bps = 50% of slot duration).
 """
 
-FAST_CONFIRM_DUE_BPS: Final = 7500
+FAST_CONFIRM_DUE_BPS: Final = Uint64(7500)
 """
 The deadline within a slot (in basis points) for achieving a fast
 confirmation.
@@ -47,7 +47,7 @@ confirmation.
 (7500 bps = 75% of slot duration).
 """
 
-VIEW_FREEZE_CUTOFF_BPS: Final = 7500
+VIEW_FREEZE_CUTOFF_BPS: Final = Uint64(7500)
 """
 The cutoff within a slot (in basis points) after which the current view is
 considered 'frozen', preventing further changes.
@@ -57,7 +57,7 @@ considered 'frozen', preventing further changes.
 
 # --- State List Length Presets ---
 
-HISTORICAL_ROOTS_LIMIT: Final = 2**18
+HISTORICAL_ROOTS_LIMIT: Final = Uint64(2**18)
 """
 The maximum number of historical block roots to store in the state.
 
@@ -65,7 +65,7 @@ With a 4-second slot, this corresponds to a history
 of approximately 12.1 days.
 """
 
-VALIDATOR_REGISTRY_LIMIT: Final = 2**12
+VALIDATOR_REGISTRY_LIMIT: Final = Uint64(2**12)
 """The maximum number of validators that can be in the registry."""
